requirements.txt
python-telegram-bot==20.*
Pillow
cairosvg
lxml
google-generativeai
numpy